docs: GET eval overview/README; split GET tests to eval_get_test.go
- Mermaid and request-flow copy for GET /evaluation
- Short shipped plan; Debug Console notes POST-only
- README Features + docs table link to Use Cases GET section
- Remove unused imports from eval_test.go after GET test split
